+++ packages/backend/src/scheduler/SchedulerTask.ts
@@ -8,6 +8,7 @@ import {
     getRequestMethod,
     GsheetsNotificationPayload,
     isChartValidationError,
+    isDashboardChartTileType,
     isDashboardValidationError,
     isEmailTarget,
     isSchedulerCsvOptions,
@@ -26,7 +27,6 @@ import {
     UploadMetricGsheetPayload,
     ValidateProjectPayload,
 } from '@lightdash/common';
-import csv from 'csvtojson';
 import { nanoid } from 'nanoid';
 import { analytics } from '../analytics/client';
 import {
@@ -50,6 +50,7 @@ import { lightdashConfig } from '../config/lightdashConfig';
 import Logger from '../logging/logger';
 import {
     csvService,
+    dashboardService,
     projectService,
     s3Service,
     schedulerService,
@@ -136,6 +137,9 @@ export const getNotificationPageData = async (
             }
             break;
         case SchedulerFormat.GSHEETS:
+            // We don't generate CSV files for Google sheets on handleNotification task,
+            // instead we directly upload the data from the row results in the uploadGsheets task
+            throw new Error("Don't fetch csv for gsheets");
         case SchedulerFormat.CSV:
             const user = await userService.getSessionByUserUuid(userUuid);
             const csvOptions = isSchedulerCsvOptions(options)
@@ -260,7 +264,7 @@ export const sendSlackNotification = async (
             throw new Error('Slack destination not found');
         }
         const { channel } = target;
-        schedulerService.logSchedulerJob({
+        await schedulerService.logSchedulerJob({
             task: 'sendSlackNotification',
             schedulerUuid,
             jobId,
@@ -355,7 +359,7 @@ export const sendSlackNotification = async (
                     pageType === LightdashPage.CHART ? 'chart' : 'dashboard',
             },
         });
-        schedulerService.logSchedulerJob({
+        await schedulerService.logSchedulerJob({
             task: 'sendSlackNotification',
             schedulerUuid,
             jobId,
@@ -378,7 +382,7 @@ export const sendSlackNotification = async (
                 type: 'slack',
             },
         });
-        schedulerService.logSchedulerJob({
+        await schedulerService.logSchedulerJob({
             task: 'sendSlackNotification',
             schedulerUuid,
             jobId,
@@ -409,7 +413,7 @@ export const testAndCompileProject = async (
             payload.createdByUserUuid,
         );
 
-        schedulerService.logSchedulerJob({
+        await schedulerService.logSchedulerJob({
             ...baseLog,
             details: { createdByUserUuid: payload.createdByUserUuid },
             status: SchedulerJobStatus.STARTED,
@@ -422,7 +426,7 @@ export const testAndCompileProject = async (
             payload.jobUuid,
         );
 
-        schedulerService.logSchedulerJob({
+        await schedulerService.logSchedulerJob({
             ...baseLog,
             details: {},
             status: SchedulerJobStatus.COMPLETED,
@@ -435,7 +439,7 @@ export const testAndCompileProject = async (
             organizationUuid: user.organizationUuid,
         });
     } catch (e) {
-        schedulerService.logSchedulerJob({
+        await schedulerService.logSchedulerJob({
             ...baseLog,
             status: SchedulerJobStatus.ERROR,
             details: { createdByUserUuid: payload.createdByUserUuid, error: e },
@@ -459,7 +463,7 @@ export const compileProject = async (
             payload.createdByUserUuid,
         );
 
-        schedulerService.logSchedulerJob({
+        await schedulerService.logSchedulerJob({
             ...baseLog,
             details: { createdByUserUuid: payload.createdByUserUuid },
             status: SchedulerJobStatus.STARTED,
@@ -471,7 +475,7 @@ export const compileProject = async (
             getRequestMethod(payload.requestMethod),
             payload.jobUuid,
         );
-        schedulerService.logSchedulerJob({
+        await schedulerService.logSchedulerJob({
             ...baseLog,
             details: {},
             status: SchedulerJobStatus.COMPLETED,
@@ -484,7 +488,7 @@ export const compileProject = async (
             organizationUuid: user.organizationUuid,
         });
     } catch (e) {
-        schedulerService.logSchedulerJob({
+        await schedulerService.logSchedulerJob({
             ...baseLog,
             status: SchedulerJobStatus.ERROR,
             details: { createdByUserUuid: payload.createdByUserUuid, error: e },
@@ -498,7 +502,7 @@ export const validateProject = async (
     scheduledTime: Date,
     payload: ValidateProjectPayload,
 ) => {
-    schedulerService.logSchedulerJob({
+    await schedulerService.logSchedulerJob({
         task: 'validateProject',
         jobId,
         scheduledTime,
@@ -546,7 +550,7 @@ export const validateProject = async (
             },
         });
 
-        schedulerService.logSchedulerJob({
+        await schedulerService.logSchedulerJob({
             task: 'validateProject',
             jobId,
             scheduledTime,
@@ -564,7 +568,7 @@ export const validateProject = async (
             },
         });
 
-        schedulerService.logSchedulerJob({
+        await schedulerService.logSchedulerJob({
             task: 'validateProject',
             jobId,
             scheduledTime,
@@ -584,20 +588,20 @@ export const downloadCsv = async (
         scheduledTime,
     };
     try {
-        schedulerService.logSchedulerJob({
+        await schedulerService.logSchedulerJob({
             ...baseLog,
             details: { createdByUserUuid: payload.userUuid },
             status: SchedulerJobStatus.STARTED,
         });
 
         const fileUrl = await csvService.downloadCsv(jobId, payload);
-        schedulerService.logSchedulerJob({
+        await schedulerService.logSchedulerJob({
             ...baseLog,
             details: { fileUrl, createdByUserUuid: payload.userUuid },
             status: SchedulerJobStatus.COMPLETED,
         });
     } catch (e) {
-        schedulerService.logSchedulerJob({
+        await schedulerService.logSchedulerJob({
             ...baseLog,
             status: SchedulerJobStatus.ERROR,
             details: { createdByUserUuid: payload.userUuid, error: e },
@@ -630,7 +634,7 @@ export const uploadGsheetFromQuery = async (
             payload.projectUuid,
             payload.exploreId,
             undefined,
-            QueryExecutionContext.CSV, // TODO chang eto gsheets
+            QueryExecutionContext.GSHEETS,
         );
         const refreshToken = await userService.getRefreshToken(
             payload.userUuid,
@@ -704,7 +708,7 @@ export const sendEmailNotification = async (
             throw new Error('Email destination not found');
         }
         const { recipient } = target;
-        schedulerService.logSchedulerJob({
+        await schedulerService.logSchedulerJob({
             task: 'sendEmailNotification',
             schedulerUuid,
             jobId,
@@ -784,7 +788,7 @@ export const sendEmailNotification = async (
                     pageType === LightdashPage.CHART ? 'chart' : 'dashboard',
             },
         });
-        schedulerService.logSchedulerJob({
+        await schedulerService.logSchedulerJob({
             task: 'sendEmailNotification',
             schedulerUuid,
             jobId,
@@ -807,7 +811,7 @@ export const sendEmailNotification = async (
                 type: 'email',
             },
         });
-        schedulerService.logSchedulerJob({
+        await schedulerService.logSchedulerJob({
             task: 'sendEmailNotification',
             schedulerUuid,
             jobId,
@@ -822,7 +826,7 @@ export const sendEmailNotification = async (
     }
 };
 
-export const sendGsheetsNotification = async (
+export const uploadGsheets = async (
     jobId: string,
     notification: GsheetsNotificationPayload,
 ) => {
@@ -853,8 +857,8 @@ export const sendGsheetsNotification = async (
             throw new Error('Missing gdriveId');
         }
 
-        schedulerService.logSchedulerJob({
-            task: 'sendGsheetsNotification',
+        await schedulerService.logSchedulerJob({
+            task: 'uploadGsheets',
             schedulerUuid,
             jobId,
             jobGroup: notification.jobGroup,
@@ -863,53 +867,74 @@ export const sendGsheetsNotification = async (
             targetType: 'gsheets',
             status: SchedulerJobStatus.STARTED,
         });
-
-        const { url, pageType, csvUrl, csvUrls } = notification.page;
+        const user = await userService.getSessionByUserUuid(
+            scheduler.createdBy,
+        );
 
         if (format !== SchedulerFormat.GSHEETS) {
             throw new Error(
                 `Unable to process format ${format} on sendGdriveNotification`,
             );
         } else if (savedChartUuid) {
-            if (csvUrl === undefined) {
-                throw new Error('Missing CSV URL');
-            }
+            const rows = await projectService.getResultsForChart(
+                user,
+                savedChartUuid,
+            );
 
-            const csvContent = await csv().fromFile(csvUrl.localPath);
             // TODO use csv().fromStream to do stream reading/writting on gsheets
             const refreshToken = await userService.getRefreshToken(
                 scheduler.createdBy,
             );
-            await googleDriveClient.appendToSheet(
-                refreshToken,
-                gdriveId,
-                csvContent,
-            );
+            await googleDriveClient.appendToSheet(refreshToken, gdriveId, rows);
         } else if (dashboardUuid) {
-            if (csvUrls === undefined) {
-                throw new Error('Missing CSV URL');
-            }
+            const dashboard = await dashboardService.getById(
+                user,
+                dashboardUuid,
+            );
+            const chartUuids = dashboard.tiles.reduce<string[]>((acc, tile) => {
+                if (
+                    isDashboardChartTileType(tile) &&
+                    tile.properties.savedChartUuid
+                ) {
+                    return [...acc, tile.properties.savedChartUuid];
+                }
+                return acc;
+            }, []);
 
             const refreshToken = await userService.getRefreshToken(
                 scheduler.createdBy,
             );
-            const googleUploadPromises = csvUrls.map(async (cu) => {
-                const csvContent = await csv().fromFile(cu.localPath);
 
+            // We want to process all charts in sequence, so we don't load all chart results in memory
+            chartUuids.reduce(async (promise, chartUuid) => {
+                await promise;
+                const rows = await projectService.getResultsForChart(
+                    user,
+                    chartUuid,
+                );
+
+                const tile = dashboard.tiles.find(
+                    (t) =>
+                        isDashboardChartTileType(t) &&
+                        t.properties.savedChartUuid === chartUuid,
+                );
+                const chartName =
+                    tile && isDashboardChartTileType(tile)
+                        ? tile.properties.chartName
+                        : undefined;
                 const tabName = await googleDriveClient.createNewTab(
                     refreshToken,
                     gdriveId,
-                    cu.filename,
+                    tile?.properties.title || chartName || chartUuid,
                 );
-                return googleDriveClient.appendToSheet(
+
+                await googleDriveClient.appendToSheet(
                     refreshToken,
                     gdriveId,
-                    csvContent,
+                    rows,
                     tabName,
                 );
-            });
-
-            Promise.all(googleUploadPromises);
+            }, Promise.resolve());
         } else {
             throw new Error('Not implemented');
         }
@@ -923,12 +948,11 @@ export const sendGsheetsNotification = async (
                 schedulerTargetId: undefined,
                 type: 'gsheets',
                 format,
-                resourceType:
-                    pageType === LightdashPage.CHART ? 'chart' : 'dashboard',
+                resourceType: savedChartUuid ? 'chart' : 'dashboard',
             },
         });
-        schedulerService.logSchedulerJob({
-            task: 'sendGsheetsNotification',
+        await schedulerService.logSchedulerJob({
+            task: 'uploadGsheets',
             schedulerUuid,
             jobId,
             jobGroup: notification.jobGroup,
@@ -949,8 +973,8 @@ export const sendGsheetsNotification = async (
                 type: 'gsheets',
             },
         });
-        schedulerService.logSchedulerJob({
-            task: 'sendGsheetsNotification',
+        await schedulerService.logSchedulerJob({
+            task: 'uploadGsheets',
             schedulerUuid,
             jobId,
             jobGroup: notification.jobGroup,
@@ -974,7 +998,7 @@ const logScheduledTarget = async (
 ) => {
     if (format === SchedulerFormat.GSHEETS) {
         await schedulerService.logSchedulerJob({
-            task: 'sendGsheetsNotification',
+            task: 'uploadGsheets',
             target: undefined,
             targetType: 'gsheets',
             jobId: targetJobId,
@@ -1033,7 +1057,7 @@ export const handleScheduledDelivery = async (
                 schedulerId: schedulerUuid,
             },
         });
-        schedulerService.logSchedulerJob({
+        await schedulerService.logSchedulerJob({
             task: 'handleScheduledDelivery',
             schedulerUuid,
             jobId,
@@ -1046,7 +1070,11 @@ export const handleScheduledDelivery = async (
             await schedulerService.schedulerModel.getSchedulerAndTargets(
                 schedulerUuid,
             );
-        const page = await getNotificationPageData(scheduler, jobId);
+
+        const page =
+            scheduler.format === SchedulerFormat.GSHEETS
+                ? undefined
+                : await getNotificationPageData(scheduler, jobId);
         const scheduledJobs =
             await schedulerClient.generateJobsForSchedulerTargets(
                 scheduledTime,
@@ -1067,7 +1095,7 @@ export const handleScheduledDelivery = async (
             );
         });
 
-        schedulerService.logSchedulerJob({
+        await schedulerService.logSchedulerJob({
             task: 'handleScheduledDelivery',
             schedulerUuid,
             jobId,
@@ -1093,7 +1121,7 @@ export const handleScheduledDelivery = async (
                 schedulerId: schedulerUuid,
             },
         });
-        schedulerService.logSchedulerJob({
+        await schedulerService.logSchedulerJob({
             task: 'handleScheduledDelivery',
             schedulerUuid,
             jobId,